requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
nltk==3.8.1
numpy==1.26.0
scikit-learn==1.3.1
pyyaml==6.0.1
spacy==3.6.1
transformers==4.34.0
torch==2.0.1
allennlp==2.10.1
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13

# Development dependencies
pytest==7.4.2
black==23.9.1
isort==5.12.0
mypy==1.5.1